Move ES6 to webpacker, get JS tests passing.
