--- app/assets/javascripts/range_limit.js
@@ -1,6 +1,0 @@
-//
-//= require 'flot/jquery.flot.js'
-//= require 'flot/jquery.flot.selection.js'
-//= require 'bootstrap-slider'
-//= require bootstrap/tooltip
-//= require blacklight_range_limit
