+++ app/javascript/figgy/auto_ingest_handler.js
