+++ app/javascript/figgy/bounding_box_selector.js
