+++ app/javascript/figgy/bulk_labeler/label_generator.js
