+++ app/javascript/figgy/collection_bag_uploader.js
