+++ app/javascript/figgy/derivative_form.js
