+++ app/javascript/figgy/file_set_form.js
