+++ app/javascript/figgy/flash.js
