+++ app/javascript/figgy/form/checklist_item.js
