+++ app/javascript/figgy/form/detect_duplicates.js
