+++ app/javascript/figgy/form/required_fields.js
