+++ app/javascript/figgy/metadata_form.js
