+++ app/javascript/figgy/modal_viewer.js
