+++ app/javascript/figgy/relationships/member_resources_table.js
