+++ app/javascript/figgy/relationships/parent_resources_table.js
