+++ app/javascript/figgy/save_and_ingest_handler.js
