+++ app/javascript/figgy/server_uploader.js
