+++ app/javascript/figgy/shift_selecting.js
