+++ app/javascript/figgy/structure_manager.js
@@ -1,5 +1,5 @@
-import {StructureParser} from "structure_parser"
-import shift_enabled_selecting from "shift_selecting"
+import {StructureParser} from "figgy/structure_parser"
+import shift_enabled_selecting from "figgy/shift_selecting"
 export default class StructureManager {
   constructor() {
     this.initialize_sortable()
