+++ app/javascript/figgy/structure_parser.js
