+++ app/javascript/figgy/universal_viewer.js
