+++ app/javascript/packs/application.js
@@ -12,13 +12,15 @@ import setupAuthLinkClipboard from '../packs/auth_link_clipboard.js'
 import AjaxSelect from '../components/ajax-select'
 import setupAjaxSelect from '../helpers/setup_ajax_select.js'
 import FileUploader from '../components/file-uploader'
+import Initializer from '../figgy/figgy_boot'
 
 Vue.use(system)
 
 // mount the filemanager app
 document.addEventListener('DOMContentLoaded', () => {
+  window.figgy = new Initializer()
   // Set CSRF token for axios requests.
-  axios.defaults.headers.common['X-CSRF-Token'] = document.querySelector('meta[name="csrf-token"]').getAttribute('content')
+  axios.defaults.headers.common['X-CSRF-Token'] = document.querySelector('meta[name="csrf-token"]') ? document.querySelector('meta[name="csrf-token"]').getAttribute('content') : undefined
   var elements = document.getElementsByClassName('lux')
   for (var i = 0; i < elements.length; i++) {
     new Vue({
