+++ app/views/layouts/blacklight/base.html.erb
@@ -14,6 +14,20 @@
     <%= csrf_meta_tag %>
     <%= stylesheet_link_tag "application", media: "all" %>
     <%= javascript_include_tag "application" %>
+    <%# There were errors when we tried to include this webpack %>
+    <%# see comments culminating with https://github.com/pulibrary/figgy/pull/3754#issuecomment-598510375 %>
+    <link href="https://cdn.jsdelivr.net/npm/select2@4.0.13/dist/css/select2.min.css" rel="stylesheet" />
+    <%= javascript_tag do %>
+      oldDefine = define;
+      define = null;
+    <% end %>
+
+    <script src="https://cdn.jsdelivr.net/npm/select2@4.0.13/dist/js/select2.js"></script>
+    <script src="//cdn.jsdelivr.net/npm/clipboard@2.0.6/dist/clipboard.js"></script>
+
+    <%= javascript_tag do %>
+      define = oldDefine;
+    <% end %>
     <%= stylesheet_pack_tag 'application' %>
     <%= javascript_pack_tag "application" %>
     <%= content_for(:head) %>
